RuslanLat-moex-main/backend/requirements.txt
aiohttp
aiohttp-apispec
aiohttp-session
alembic
asyncpg
cryptography
lxml
marshmallow
PyYAML
SQLAlchemy
